# src/main/resources/ob-cnf-account-request-data-api-swagger.yaml
#Open Banking Account Request Data Enterprise API
swagger: '2.0'
info:
  title: Open Banking Account Request Data Enterprise API
  description: Swagger specification for Account Request Data Enterprise API
  version: v0.1
# the domain of the service
host: api.lbg.com
basePath: /account-request-data/v1
schemes:
  - https
produces:
  - application/json
securityDefinitions:
  BasicAuth:
    type: basic
    description: Basic authentication
security:
  - BasicAuth: []
paths:
  /accounts-requests:
    post:
      summary: Create an account request
      description: Create an account request
      operationId: CreateAccountRequest
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: AccountRequest
          description: Create an Account Request
          in: body
          required: true
          schema:
            $ref: '#/definitions/AccountRequest'
        - $ref: '#/parameters/xLbgTxnCorrelationId'
        - $ref: '#/parameters/xLbgClientId'
        - $ref: '#/parameters/xFapiFinancialId'
        - $ref: '#/parameters/xFapiInteractionId'
        - name: x-lbg-internal-user-role
          description: Internal User Role Header, should be SYSTEM for create
          in: header
          required: true
          type: string
          enum:
            - SYSTEM

      responses:
        201:
          description: Account Request resource successfully created
          schema:
            $ref: '#/definitions/AccountRequestResponse'
          headers:
            x-fapi-interaction-id:
              type: string
              description: An RFC4122 UID used as a correlation id.
            Strict-Transport-Security:
                default: max-age=0
                description: HTTPS strict transport security header
                type: string
            Cache-Control:
              default: max-age=0,no-cache,no-store
              type: string
              description: Describes how long this response can be cached
            X-Frame-Options:
              default: DENY
              type: string
              description: Prevent this request from being loaded in any iframes
            X-Content-Type-Options:
              type: string
              description: >-
                Ensures each page has a content type and prevents browsers from
                doing MIME type sniffing
              default: nosniff
        400:
          $ref: "#/responses/Standard400ErrorResponse"
        401:
          $ref: "#/responses/Standard401ErrorResponse"
        403:
          $ref: "#/responses/Standard403ErrorResponse"
        415:
          $ref: "#/responses/Standard415ErrorResponse"
        500:
          $ref: "#/responses/Standard500ErrorResponse"
        503:
          $ref: "#/responses/Standard503ErrorResponse"
    get:
      summary: Get an account request for provided query parameters
      description: Get an account request for provided query parameters
      operationId: GetAccountRequestByQueryParams
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/AccountRequestIdQueryParam'
        - $ref: '#/parameters/ClientId'
        - $ref: '#/parameters/xLbgTxnCorrelationId'
        - name: x-lbg-internal-user-role
          description: Internal User Role Header, should be CUSTOMER for get with query
          in: header
          required: true
          type: string
          enum:
            - CUSTOMER
      responses:
        200:
          description: Account Request resource successfully retrieved
          schema:
            $ref: '#/definitions/AccountRequestResponse'
          headers:
            Strict-Transport-Security:
                default: max-age=0
                description: HTTPS strict transport security header
                type: string
            Cache-Control:
              default: max-age=0,no-cache,no-store
              type: string
              description: Describes how long this response can be cached
            X-Frame-Options:
              default: DENY
              type: string
              description: Prevent this request from being loaded in any iframes
            X-Content-Type-Options:
              type: string
              description: >-
                Ensures each page has a content type and prevents browsers from
                doing MIME type sniffing
              default: nosniff
        400:
          $ref: "#/responses/Standard400ErrorResponse"
        401:
          $ref: "#/responses/Standard401ErrorResponse"
        403:
          $ref: "#/responses/Standard403ErrorResponse"
        500:
          $ref: "#/responses/Standard500ErrorResponse"
        503:
          $ref: "#/responses/Standard503ErrorResponse"
  /accounts-requests/{AccountRequestId}:
    get:
      summary: Get an account request for unique account request identifier
      description: Get an account request for unique account request identifier
      operationId: GetAccountRequestByRequestId
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/AccountRequestId'
        - $ref: '#/parameters/xLbgTxnCorrelationId'
        - $ref: '#/parameters/xFapiInteractionId'
        - name: x-lbg-internal-user-role
          description: Internal User Role Header, should be SYSTEM for GET with path
          in: header
          required: true
          type: string
          enum:
            - SYSTEM
        - name: x-fapi-financial-id
          description: The unique id of the ASPSP to which the request is issued. The unique id will be issued by OB.
          in: header
          required: false
          type: string
        - name: x-lbg-client-id
          description: Client Id Header
          in: header
          required: false
          type: string
      responses:
        200:
          description: Account Request resource successfully retrieved
          schema:
            $ref: '#/definitions/AccountRequestResponse'
          headers:
            x-fapi-interaction-id:
              type: string
              description: An RFC4122 UID used as a correlation id.
            Strict-Transport-Security:
                default: max-age=0
                description: HTTPS strict transport security header
                type: string
            Cache-Control:
              default: max-age=0,no-cache,no-store
              type: string
              description: Describes how long this response can be cached
            X-Frame-Options:
              default: DENY
              type: string
              description: Prevent this request from being loaded in any iframes
            X-Content-Type-Options:
              type: string
              description: >-
                Ensures each page has a content type and prevents browsers from
                doing MIME type sniffing
              default: nosniff
        400:
          $ref: "#/responses/Standard400ErrorResponse"
        401:
          $ref: "#/responses/Standard401ErrorResponse"
        403:
          $ref: "#/responses/Standard403ErrorResponse"
        500:
          $ref: "#/responses/Standard500ErrorResponse"
        503:
          $ref: "#/responses/Standard503ErrorResponse"
    delete:
      summary: Delete an account request
      description: AISP can delete a previosly created account request (whether it was consented to or not). The PSU may want to remove their consent with AISP instead of revoking authorisation with the ASPSP
      operationId: DeleteAccountRequest
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/AccountRequestId'
        - $ref: '#/parameters/xLbgTxnCorrelationId'
        - $ref: '#/parameters/xFapiFinancialId'
        - $ref: '#/parameters/xLbgClientId'
        - $ref: '#/parameters/xFapiInteractionId'
        - name: x-lbg-internal-user-role
          description: Internal User Role Header, should be SYSTEM for delete
          in: header
          required: true
          type: string
          enum:
            - SYSTEM
      responses:
        204:
          description: No Content if Account Request resource successfully deleted
          headers:
            x-fapi-interaction-id:
              type: string
              description: An RFC4122 UID used as a correlation id.
            Strict-Transport-Security:
                default: max-age=0
                description: HTTPS strict transport security header
                type: string
            Cache-Control:
              default: max-age=0,no-cache,no-store
              type: string
              description: Describes how long this response can be cached
            X-Frame-Options:
              default: DENY
              type: string
              description: Prevent this request from being loaded in any iframes
            X-Content-Type-Options:
              type: string
              description: >-
                Ensures each page has a content type and prevents browsers from
                doing MIME type sniffing
              default: nosniff
        400:
          $ref: "#/responses/Standard400ErrorResponse"
        401:
          $ref: "#/responses/Standard401ErrorResponse"
        403:
          $ref: "#/responses/Standard403ErrorResponse"
        500:
          $ref: "#/responses/Standard500ErrorResponse"
  /accounts-requests/{AccountRequestId}/status:
    put:
      summary: Update status of account request and optionally set account ids, if status is Authorised
      description: Update status of account request and optionally set account ids, if status is Authorised
      operationId: UpdateAccountRequestStatus
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/AccountRequestId'
        - $ref: '#/parameters/xLbgTxnCorrelationId'
        - $ref: '#/parameters/xLbgClientId'
        - name: x-lbg-internal-user-role
          description: Internal User Role Header, should be CUSTOMER for update
          in: header
          required: true
          type: string
          enum:
            - CUSTOMER
        - name: AccountRequestStatusInputData
          description: Account request status input data
          in: body
          required: true
          schema:
            $ref: '#/definitions/UpdateAccountRequestStatusInputData'
      responses:
        200:
          description: Account Request resource successfully created
          schema:
            $ref: '#/definitions/UpdatedAccountRequestStatusOutputData'
          headers:
            Strict-Transport-Security:
              default: max-age=0
              description: HTTPS strict transport security header
              type: string
            Cache-Control:
              default: max-age=0,no-cache,no-store
              type: string
              description: Describes how long this response can be cached
            X-Frame-Options:
              default: DENY
              type: string
              description: Prevent this request from being loaded in any iframes
            X-Content-Type-Options:
              type: string
              description: >-
                Ensures each page has a content type and prevents browsers from
                doing MIME type sniffing
              default: nosniff
        400:
          $ref: "#/responses/Standard400ErrorResponse"
        401:
          $ref: "#/responses/Standard401ErrorResponse"
        403:
          $ref: "#/responses/Standard403ErrorResponse"
        415:
          $ref: "#/responses/Standard415ErrorResponse"
        500:
          $ref: "#/responses/Standard500ErrorResponse"
        503:
          $ref: "#/responses/Standard503ErrorResponse"
definitions:
  UpdateAccountRequestStatusInputData:
    type: object
    properties:
      Status:
        $ref: '#/definitions/ApplicableAuthorisationStatus'
      EntitlementId:
        type: integer
        description: the entitlement id that will be passed in by the entitlement API as a reference to the entitlement against a request to authorise the account request.
    required:
      - Status
  UpdatedAccountRequestStatusOutputData:
    type: object
    properties:
      AccountRequestId:
        type: string
      UpdatedStatus:
        $ref: '#/definitions/ApplicableAuthorisationStatus'
      UpdatedAtTimestamp:
        $ref: '#/definitions/Timestamp'
    required:
      - AccountRequestId
      - UpdatedStatus
      - UpdatedAtTimestamp
  Timestamp:
    type: string
    format: date-time
    description: ISO 8601 compliant Timestamp
  AccountRequest:
    title: Account Request POST request
    description: Allows setup of an account access request
    type: object
    properties:
      Data:
        $ref: '#/definitions/Data'
      Risk:
        $ref: '#/definitions/Risk'
    additionalProperties: false
    required:
      - Data
      - Risk
  Data:
    title: Account Request Data
    type: object
    properties:
      Permissions:
        title: Permissions list
        description: Specifies the Open Banking account request types. This is a list of the data clusters being consented by the PSU, and requested for authorisation with the ASPSP.
        type: array
        items:
          type: string
          properties:
            Code:
              $ref: '#/definitions/PermissionCode'
          minProperties: 1
          additionalProperties: false
      ExpirationDateTime:
        $ref: '#/definitions/ExpirationDateTime'
      TransactionFromDateTime:
        $ref: '#/definitions/TransactionFromDateTime'
      TransactionToDateTime:
        $ref: '#/definitions/TransactionToDateTime'
    additionalProperties: false
    required:
      - Permissions

  Risk:
    title: Account Request Risk object
    type: object
    additionalProperties: false

  AccountRequestResponse:
    title: Account Request GET response
    description: Account request GET response after successful account request creation
    type: object
    properties:
      Data:
        $ref: '#/definitions/ResponseData'
      Risk:
        $ref: '#/definitions/Risk'
    additionalProperties: false
    required:
      - Data
      - Risk
  ResponseData:
    title: Account Request Data
    type: object
    properties:
      AccountRequestId:
        $ref: '#/definitions/AccountRequestId'
      Status:
        $ref: '#/definitions/Status'
      Permissions:
        $ref: '#/definitions/Permissions'
      CreationDateTime:
        $ref: '#/definitions/CreationDateTime'
      ExpirationDateTime:
        $ref: '#/definitions/ExpirationDateTime'
      TransactionFromDateTime:
        $ref: '#/definitions/TransactionFromDateTime'
      TransactionToDateTime:
        $ref: '#/definitions/TransactionToDateTime'
    additionalProperties: false
    required:
      - Permissions
  AccountRequestId:
    title: AccountRequestId
    description: Unique identification as assigned to identify the account request resource
    type: string
    minLength: 1
    maxLength: 40
  Permissions:
    title: Permissions list
    description: Specifies the Open Banking account request types. This is a list of the data clusters being consented by the PSU, and requested for authorisation with the ASPSP.
    type: array
    items:
      $ref: '#/definitions/Permission'
    minProperties: 1
    additionalProperties: false
  PermissionCode:
    title: PermissionCode
    description: A unique permission code.
    type: string
    enum:
      - ReadAccountsBasic
      - ReadAccountsDetail
      - ReadBalances
      - ReadBeneficiariesBasic
      - ReadBeneficiariesDetail
      - ReadDirectDebits
      - ReadProducts
      - ReadStandingOrdersBasic
      - ReadStandingOrdersDetail
      - ReadTransactionsBasic
      - ReadTransactionsCredits
      - ReadTransactionsDebits
      - ReadTransactionsDetail
  Permission:
    type: object
    properties:
      Code:
        $ref: '#/definitions/PermissionCode'
      Description:
        type: string
      SectionHeader:
        type: string
  ExpirationDateTime:
    title: ExpirationDateTime
    description: Specified date and time the permissions will expire. If this is not populated, the permissions will be open ended.
    type: string
    format: date-time
  CreationDateTime:
    title: CreationDateTime
    description: The creation date and time of the Account request.
    type: string
    format: date-time
  ApplicableAuthorisationStatus:
    title: ApplicableAuthorisationStatus
    description: Specifies the status of the account request resource that can be updated after successful/unsuccessful authorisation
    type: string
    enum:
      - Authorised
      - Rejected
  Status:
    title: Status
    description: Specifies the status of the account request resource in code form.
    type: string
    enum:
      - AwaitingAuthorisation
      - Authorised
      - Rejected
      - Revoked
  TransactionFromDateTime:
    title: TransactionFromDateTime
    description: Specified start date and time for the transaction query period. If this is not populated, the request will be open ended.
    type: string
    format: date-time
  TransactionToDateTime:
    title: TransactionToDateTime
    description: Specified end date and time for the transaction query period. If this is not populated, the request will be open ended.
    type: string
    format: date-time
  ErrorObject:
    type: object
    properties:
      code:
        type: string
      message:
        type: string
      statusCode:
        type: number
  Error:
    type: object
    properties:
      error:
        $ref: '#/definitions/ErrorObject'
parameters:
  AccountRequestId:
    name: AccountRequestId
    description: Unique identification as assigned by the ASPSP to uniquely identify the account request resource.
    type: string
    in: path
    required: true
  AccountRequestIdQueryParam:
    name: AccountRequestId
    description: Unique identification as assigned by the ASPSP to uniquely identify the account request resource.
    type: string
    in: query
    required: true
  ClientId:
    name: ClientId
    description: Unique identification of TPP assigned by LBG
    type: string
    in: query
    required: true
  xLbgTxnCorrelationId:
    name: x-lbg-txn-correlation-id
    description: Transaction Correlation Id Header
    in: header
    required: true
    type: string
  xLbgInternalUserRole:
    name: x-lbg-internal-user-role
    description: Internal User Role Header
    in: header
    required: true
    type: string
  xFapiFinancialId:
    name: x-fapi-financial-id
    description: The unique id of the ASPSP to which the request is issued. The unique id will be issued by OB.
    in: header
    required: true
    type: string
  xFapiInteractionId:
    name: x-fapi-interaction-id
    description: An RFC4122 UID used as a correlation id.
    in: header
    required: false
    type: string
  xLbgClientId:
    name: x-lbg-client-id
    description: Client Id Header
    in: header
    required: true
    type: string
responses:
  Standard400ErrorResponse:
    description: Bad request
    schema:
      $ref: "#/definitions/Error"
  Standard401ErrorResponse:
    description: Unauthorised
    schema:
      $ref: "#/definitions/Error"
  Standard403ErrorResponse:
    description: Forbidden
    schema:
      $ref: "#/definitions/Error"
  Standard415ErrorResponse:
    description: Unsupported media type
    schema:
      $ref: "#/definitions/Error"
  Standard500ErrorResponse:
    description: Unexpected server error
    schema:
      $ref: "#/definitions/Error"
  Standard503ErrorResponse:
    description: Service unavailable error
    schema:
      $ref: "#/definitions/Error"